added support for point, rectangular, and directional lights
